fix(karma): Corrected development dependencies
Although the peer dependency was correct, the dev dependency to karma still referenced a very old version, which did not yet support async framework initialization. This lead to the unit tests failing due to race conditions, depending on the system. In addition, the server startup promise was not returned and thus not respected by the async framework initialization.